Forest Attributes in Washington State

The U.S. Department of Agriculture, Forest Service, Forest Inventory and Analysis (FIA) Program provides all sorts of estimates of forest attributes for uses in research, legislation, and land management. The FIA uses a set of criteria to classify a plot of land as "forested" or "non-forested," and that classification is a central data point in many decision-making contexts. A small subset of plots in Washington State are sampled and assessed "on-the-ground" as forested or non-forested, but the FIA has access to remotely sensed data for all land in the state. Practitioners can develop a model on the more easily-accessible remotely sensed data to predict whether a plot is forested or non-forested.

Usage

forested

Format

A data frame with 7,107 rows and 19 columns:

forested

Whether the plot is classified as "forested" or not, as a factor with levels "Yes" and "No".

year

Year when the plot was classified "on-the-ground" as forested or not. The remaining, remotely-sensed variables are measured at different times or averaged over multiple years.

elevation

Elevation, in meters.

eastness

Transformed aspect degrees to eastness (-100 to 100).

northness

Transformed aspect degrees to northness (-100 to 100).

roughness

Degree of irregularity of the plot.

tree_no_tree

LANDFIRE tree/non-tree lifeform mask, as a factor with levels "Tree" and "No tree".

dew_temp

Mean annual dewpoint temperature (1991-2020), in degrees Celsius.

precip_annual

Mean annual precipitation (1991-2020), in mm × 100.

temp_annual_mean

Mean annual temperature (1991-2020), in degrees Celsius.

temp_annual_min

Mean annual minimum temperature (1991-2020), in degrees Celsius.

temp_annual_max

Mean annual maximum temperature (1991-2020), in degrees Celsius.

temp_january_min

Mean minimum temperature in January (1991-2020), in degrees Celsius.

vapor_min, vapor_max

Minimum and maximum annual vapor pressure deficit (1991-2020), in Pa x 100.

canopy_cover

Analytical Tree Canopy Cover, as a percent.

lon, lat

The longitude and latitude of the center of the plot with a slight perturbation.

land_type

Land cover type from European Space Agency (ESA) 2020 WorldCover global land cover product, as a factor with levels "Tree", "Non-tree vegetation", and "Barren".

Source

For more information on the source data, see Table 1 in:

White, Grayson W.; Yamamoto, Josh K.; Elsyad, Dinan H.; Schmitt, Julian F.; Korsgaard, Niels H.; Hu, Jie Kate; Gaines III, George C.; Frescino Tracey S.; McConville, Kelly S. (2024). Small area estimation of forest biomass via a two-stage model for continuous zero-inflated data. Forthcoming: arXiv 2402.03263 (ver. 2.0).

For more on data definitions:

Wieczorek, Jerzy A.; White, Grayson W.; Cody, Zachariah W.; Tan, Emily X.; Chistolini, Jacqueline O.; McConville, Kelly S.; Frescino, Tracey S.; Moisen, Gretchen G. (2024). Assessing small area estimates via artificial populations from KBAABB: a kNN-based approximation to ABB. Forthcoming: arXiv 2306.15607 (ver. 2.0.

Source data pre-preprocessed using the FIESTA R Package (GPL-3):

Frescino, Tracey S.; Moisen, Gretchen G.; Patterson, Paul L.; Toney, Chris; White, Grayson W. (2023). FIESTA: A forest inventory estimation and analysis R package. Ecography 2023: e06428 (ver. 1.0).

Examples

str(forested)
#> tibble [7,107 × 19] (S3: tbl_df/tbl/data.frame)
#>  $ forested        : Factor w/ 2 levels "Yes","No": 1 1 2 1 1 1 1 1 1 1 ...
#>  $ year            : num [1:7107] 2005 2005 2005 2005 2005 ...
#>  $ elevation       : num [1:7107] 881 113 164 299 806 736 636 224 52 2240 ...
#>  $ eastness        : num [1:7107] 90 -25 -84 93 47 -27 -48 -65 -62 -67 ...
#>  $ northness       : num [1:7107] 43 96 53 34 -88 -96 87 -75 78 -74 ...
#>  $ roughness       : num [1:7107] 63 30 13 6 35 53 3 9 42 99 ...
#>  $ tree_no_tree    : Factor w/ 2 levels "Tree","No tree": 1 1 1 2 1 1 2 1 1 2 ...
#>  $ dew_temp        : num [1:7107] 0.04 6.4 6.06 4.43 1.06 1.35 1.42 6.39 6.5 -5.63 ...
#>  $ precip_annual   : num [1:7107] 466 1710 1297 2545 609 ...
#>  $ temp_annual_mean: num [1:7107] 6.42 10.64 10.07 9.86 7.72 ...
#>  $ temp_annual_min : num [1:7107] -8.32 1.4 0.19 -1.2 -5.98 ...
#>  $ temp_annual_max : num [1:7107] 12.9 15.8 14.4 15.8 13.8 ...
#>  $ temp_january_min: num [1:7107] -0.08 5.44 5.72 3.95 1.6 1.12 0.99 5.54 6.2 -4.54 ...
#>  $ vapor_min       : num [1:7107] 78 34 49 67 114 67 67 31 60 79 ...
#>  $ vapor_max       : num [1:7107] 1194 938 754 1164 1254 ...
#>  $ canopy_cover    : num [1:7107] 50 79 47 42 59 36 14 27 82 12 ...
#>  $ lon             : num [1:7107] -119 -123 -122 -122 -118 ...
#>  $ lat             : num [1:7107] 48.7 47.1 48.8 45.8 48.1 ...
#>  $ land_type       : Factor w/ 3 levels "Barren","Non-tree vegetation",..: 3 3 3 3 3 3 2 2 3 2 ...

head(forested)
#> # A tibble: 6 × 19
#>   forested  year elevation eastness northness roughness tree_no_tree dew_temp
#>   <fct>    <dbl>     <dbl>    <dbl>     <dbl>     <dbl> <fct>           <dbl>
#> 1 Yes       2005       881       90        43        63 Tree             0.04
#> 2 Yes       2005       113      -25        96        30 Tree             6.4 
#> 3 No        2005       164      -84        53        13 Tree             6.06
#> 4 Yes       2005       299       93        34         6 No tree          4.43
#> 5 Yes       2005       806       47       -88        35 Tree             1.06
#> 6 Yes       2005       736      -27       -96        53 Tree             1.35
#> # ℹ 11 more variables: precip_annual <dbl>, temp_annual_mean <dbl>,
#> #   temp_annual_min <dbl>, temp_annual_max <dbl>, temp_january_min <dbl>,
#> #   vapor_min <dbl>, vapor_max <dbl>, canopy_cover <dbl>, lon <dbl>, lat <dbl>,
#> #   land_type <fct>